+++ sdl2/pdcscrn.c
@@ -419,7 +419,7 @@ int PDC_color_content(int color, int *red, int *green, int *blue)
     return OK;
 }
 
-int PDC_init_color(short color, short red, short green, short blue)
+int PDC_init_color(int color, int red, int green, int blue)
 {
     pdc_color[color].r = DIVROUND(red * 255, 1000);
     pdc_color[color].g = DIVROUND(green * 255, 1000);
